refactor: rename styles/ dir to theme/
