+++ libs/spark/src/theme/overrides.ts
