+++ libs/spark/src/theme/palette.ts
